requirements.txt
amqp==2.5.2
asgiref==3.2.7
attrs==19.3.0
autobahn==20.4.3
Automat==20.2.0
Babel==2.8.0
billiard==3.6.3.0
boto==2.49.0
boto3==1.13.16
botocore==1.16.16
celery==4.4.2
certifi==2020.4.5.1
cffi==1.14.0
chardet==3.0.4
constantly==15.1.0
cryptography==2.9.2
daphne==2.5.0
defusedxml==0.6.0
diff-match-patch==20181111
dj-database-url==0.5.0
Django==3.0.6
django-allauth==0.41.0
django-autocomplete-light==3.5.1
django-celery-results==1.2.1
django-crispy-forms==1.9.1
django-elasticsearch-dsl==7.1.1
django-environ==0.4.5
django-import-export==2.1.0
django-phonenumber-field==4.0.0
django-storages==1.9.1
djangorestframework==3.11.0
docutils==0.15.2
elasticsearch==7.7.0
elasticsearch-dsl==7.2.0
et-xmlfile==1.0.1
googlemaps==4.4.0
hyperlink==19.0.0
idna==2.9
importlib-metadata==1.6.0
incremental==17.5.0
jdcal==1.4.1
jmespath==0.10.0
kombu==4.6.8
MarkupPy==1.14
oauthlib==3.1.0
odfpy==1.4.1
openpyxl==3.0.3
phonenumbers==8.12.4
phonenumberslite==8.12.4
Pillow==7.1.2
psycopg2-binary==2.8.5
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
PyHamcrest==2.0.2
pyOpenSSL==19.1.0
python-dateutil==2.8.1
python3-openid==3.1.0
pytz==2020.1
PyYAML==5.3.1
redis==3.5.2
requests==2.23.0
requests-oauthlib==1.3.0
s3transfer==0.3.3
sentry-sdk==0.14.4
service-identity==18.1.0
six==1.15.0
sqlparse==0.3.1
tablib==2.0.0
Twisted==20.3.0
txaio==20.4.1
urllib3==1.25.9
vine==1.3.0
whitenoise==5.1.0
xlrd==1.2.0
xlwt==1.3.0
zipp==3.1.0
zope.interface==5.1.0
